update function in luggage

+++ LuggageEye/demo/src/main/java/edu/depaul/cdm/se452/demo/Luggage/Luggage.java
@@ -21,8 +21,8 @@ public class Luggage {
 
     }
 
-    public int calculateAmount(int weight, float price) {
-        int amount = weight * price;
+    public float calculateAmount(int weight, float price) {
+        float amount = weight * price;
         
         return amount;
